fix(codegraph-parser): support ABI-15 grammars via tree-sitter 0.26
Also swaps deprecated criterion::black_box for std::hint::black_box.
Bumps crates/knowledge to tree-sitter 0.26 and points repomix-config /
repomix-core at the vendored third_party/upstream/repomix-rs crates
(the crates.io 2.0.0 releases pin tree-sitter 0.24, which conflicts on
the native links key), matching the fix already validated on the task
branch. Adapts Node::child callers to the u32 index API.

Unblocks #130 #134 #136 #137 #138 #139; develop needs the same fix for #131.

+++ crates/external/codegraph-parser/src/languages/go.rs
@@ -131,7 +131,7 @@ impl<'a> GoCollector<'a> {
             "import_declaration" => {
                 // Handle both single imports and import blocks
                 for i in 0..node.child_count() {
-                    if let Some(child) = node.child(i) {
+                    if let Some(child) = node.child(i as u32) {
                         if child.kind() == "import_spec" || child.kind() == "import_spec_list" {
                             self.extract_import_spec(&child, &ctx);
                         }
@@ -244,7 +244,7 @@ impl<'a> GoCollector<'a> {
             // Go Type declarations (struct, interface)
             "type_declaration" => {
                 for i in 0..node.child_count() {
-                    if let Some(spec) = node.child(i) {
+                    if let Some(spec) = node.child(i as u32) {
                         if spec.kind() == "type_spec" {
                             self.extract_type_spec(&spec, &mut ctx);
                         }
